# backend/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/team/register": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Register User",
        "operationId": "register_user_team_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/team/create": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Create Team",
        "operationId": "create_team_team_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/team/join": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Join Team",
        "operationId": "join_team_team_join_post",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/team/assign-color": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Assign Color To User",
        "description": "Assign a unique color to a user within their team.",
        "operationId": "assign_color_to_user_team_assign_color_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignColorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColorAssignmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/team/{team_id}/validate-colors": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "Validate Team Colors",
        "description": "Validate that all players in a team have unique colors.",
        "operationId": "validate_team_colors_team__team_id__validate_colors_get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamColorValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/team/{team_id}/resolve-conflicts": {
      "post": {
        "tags": [
          "team"
        ],
        "summary": "Resolve Color Conflicts",
        "description": "Resolve any color conflicts in a team by reassigning colors.",
        "operationId": "resolve_color_conflicts_team__team_id__resolve_conflicts_post",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColorConflictResolutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/team/{team_id}/available-colors": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "Get Available Colors",
        "description": "Get available and used colors for a team.",
        "operationId": "get_available_colors_team__team_id__available_colors_get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableColorsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignColorRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "ID of the user to assign color to"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id",
            "description": "ID of the team the user belongs to"
          },
          "preferred_color": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Preferred Color",
            "description": "Preferred color (optional)"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "team_id"
        ],
        "title": "AssignColorRequest",
        "description": "Assign Color Request: Request to assign a color to a user"
      },
      "AvailableColorsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "AvailableColorsResponse",
        "description": "Available Colors Response: Response for available colors"
      },
...
      "ColorAssignmentResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ColorAssignmentResponse",
        "description": "Color Assignment Response: Response containing color assignment result"
      },
      "ColorConflictResolutionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ColorConflictResolutionResponse",
        "description": "Color Conflict Resolution Response: Response for color conflict resolution"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TeamColorValidationResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "TeamColorValidationResponse",
        "description": "Team Color Validation Response: Response for team color validation"
      },
      "TeamCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Name for the new team"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TeamCreate",
        "description": "Team Create: Request to create a new team"
      },
      "TeamOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Team ID"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Team name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "TeamOut",
        "description": "Team Out: Team response model"
      },
...
      "UserCreate": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Username for the new user"
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "UserCreate",
        "description": "User Create: Request to create a new user"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "User ID"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Username"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id",
            "description": "Team ID"
          },
          "points": {
            "type": "integer",
            "title": "Points",
            "description": "Current points"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color",
            "description": "Assigned color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "points"
        ],
        "title": "UserOut",
        "description": "User Out: User response model"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}